turned line pick example into a line select example
